+++ src/typeguard/_checkers.py
@@ -556,6 +556,11 @@ def check_type_internal(value: Any, annotation: Any, memo: TypeCheckMemo) -> Non
 
     if origin_type is not None:
         args = get_args(annotation)
+
+        # Compatibility hack to distinguish between unparametrized and empty tuple
+        # (tuple[()]), necessary due to https://github.com/python/cpython/issues/91137
+        if origin_type in (tuple, Tuple) and annotation is not Tuple and not args:
+            args = ((),)
     else:
         origin_type = annotation
         args = ()